Fix footer not sticking to bottom in IE
Internet Explorer does not properly calculate the height of flexed
children of containers with min-height if they don't have an explicit
height property. This caused the content div to have a height of 0,
meaning the footer would appear at the top of the page. Changing
min-height to height seems to have no side effects, and fixes this
issue.

This was only tested in the latest version of Chrome (57) and the last
version of Internet Explorer (11), both on Windows. It is expected to
work in all recent browsers (after 2013).

Sources:

https://github.com/philipwalton/flexbugs/blob/c152693862bd0b5a1964eecd13050fb9bf0fe72d/README.md#3-min-height-on-a-flex-container-wont-apply-to-its-flex-items

https://connect.microsoft.com/IE/feedback/details/802625/min-height-and-flexbox-flex-direction-column-dont-work-together-in-ie-10-11-preview